# doc/api.yaml
openapi: 3.0.3
servers:
  - url: "http://localhost:3000"
info:
  title: Wasa Photo
  description: |-
    Wasa Photo is a photo sharing platform. It allows users to upload photo of their best moment.

  version: 1.0.0
...
paths:
...
  /profiles/{profile_user_id}/followings:
    parameters:
      - name: profile_user_id
        in: path
        description: The user Id of the followed user
        schema: { $ref: "#/components/schemas/UserID" }
        required: true

    post:
      tags: ["User"]
      operationId: followUser
      summary: Follow user
      description: |-
        Follow the user with the target_user_id.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: |-
            Follow user successful. <br>
            Return the user ID and username of the followed user.
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/OK" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    delete:
      tags: ["User"]
      operationId: unfollowUser
      summary: Unfollow user
      description: |-
        Unfollow the user with the given user ID and delete it from the follower list.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: |-
            Unfollow user successful. <br>
            Return the user ID and the username of the unfollowed user.
          content:
            text/plain:
              schema: { $ref: "#/components/schemas/OK" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    get:
      tags: ["User"]
      operationId: getMyFollowings
      summary: Get my followings
      description: |-
        Get the followings of the user with the given user ID.
        Return a list of users. See User for more details.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Get followings successful
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsersList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profiles/{profile_user_id}/followers:
    parameters:
      - name: profile_user_id
        in: path
        description: The user Id of the owner of the post.
        schema: { $ref: "#/components/schemas/UserID" }
        required: true
      - { $ref: "#/components/parameters/Offset" }
      - { $ref: "#/components/parameters/Limit" }

    get:
      tags: ["User"]
      operationId: getMyFollowers
      summary: Get my followers
      description: |-
        Get the followers of the user with the given user ID.
        Return a list of users. See User for more details.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Get followers successful
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsersList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /profile/{profile_user_id}/bans:
    parameters:
      - name: profile_user_id
        in: path
        description: The user Id of the banned user
        schema: { $ref: "#/components/schemas/UserID" }
        required: true

    post:
      tags: ["User"]
      operationId: banUser
      summary: Ban user
      description: |-
        Ban the user with the given user ID
      security:
        - bearerAuth: []
      responses:
        "200":
          description: |-
            Ban user successful. <br>
            Return the user of the banned user. See User for more details.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    delete:
      tags: ["User"]
      operationId: unbanUser
      summary: Unban user
      description: |-
        Unban the user with the given user ID
      security:
        - bearerAuth: []
      responses:
        "200":
          description: |-
            Unban user successful
          content:
            text/plain:
              schema: { $ref: "#/components/schemas/OK" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    get:
      tags: ["User"]
      operationId: getMyBans
      summary: Get my bans
      description: |-
        Get the bans of the user with the given user ID.
        Return a list of users. See User for more details.
      security:
        - bearerAuth: []

      parameters:
        - { $ref: "#/components/parameters/Offset" }
        - { $ref: "#/components/parameters/Limit" }

      responses:
        "200":
          description: Get bans successful
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UsersList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /profiles/{profile_user_id}/posts/{post_id}:
    parameters:
      - name: profile_user_id
        in: path
        description: |-
          The user Id of the owner of the post.
        schema: { $ref: "#/components/schemas/UserID" }
        required: true
      - name: post_id
        in: path
        description: The post Id of the post to delete
        schema: { $ref: "#/components/schemas/PostID" }
        required: true

    delete:
      tags: ["Post"]
      operationId: deletePhoto
      summary: Delete post
      description: |-
        Delete the specified post by post id from the storage server, 
        delete the post from user profile and delete all like and comment related to it.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Delete photo and post successful
          content:
            text/plain:
              schema: { $ref: "#/components/schemas/OK" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /profiles/{profile_user_id}/posts/{post_id}/comments:
    parameters:
      - name: profile_user_id
        in: path
        description: The user Id of the owner of the post.
        schema: { $ref: "#/components/schemas/UserID" }
        required: true
      - name: post_id
        in: path
        description: The post Id of the post to comment
        schema: { $ref: "#/components/schemas/PostID" }
        required: true

    post:
      tags: ["Post"]
      operationId: commentPhoto
      summary: Comment post
      description: |-
        Comment post
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Comment" }
      responses:
        "201":
          description: Comment created successful
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    get:
      tags: ["Post"]
      operationId: getComments
      summary: Get a subset or all comments of a post
      description: |-
        Get a subset or all comments of a post. 
        It's possible to specify a range and limit of comments to get by using query parameters.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Get comments successful
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommentsList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  schemas:
    UserID:
      title: User id
      description: Numeric id of the user
      type: integer
      example: 10
      minimum: 1
    Username:
      title: Username
      description: Username of the user, not accepted upper case characters
      type: string
      pattern: ^[a-z][a-z0-9_]{2,13}$
      example: "haru_22"
      minLength: 3
      maxLength: 13
    User:
      title: User
      description: User
      type: object
      properties:
        user_id: { $ref: "#/components/schemas/UserID" }
        username: { $ref: "#/components/schemas/Username" }
        propic_url: { $ref: "#/components/schemas/ImgURL" }
...
    PostID:
      title: PostID
      description: |-
        The post id is used to identify a post.
        The post id is generated by the server and it's for each user.
      type: integer
      example: 100
      minimum: 1
    CommentID:
      title: CommentID
      description: |-
        The numeric id is generated by the server and it's unique for each comments in a post.
      type: integer
      example: 15
      minimum: 1
    Comment:
      title: Comment
      description: Comment of the post
      type: object
      properties:
        user: { $ref: "#/components/schemas/User" }
        comment_id: { $ref: "#/components/schemas/CommentID" }
        comment:
          type: string
          description: Text of the comment
          example: "Bella foto!"
          pattern: '^[^\/\\]{1,100}$'
          minLength: 1
          maxLength: 100
        timestamp: { $ref: "#/components/schemas/Timestamp" }
    CommentsList:
      title: Comment List
      description: All comment of the post
      type: array
      items: { $ref: "#/components/schemas/Comment" }
      minItems: 0
      maxItems: 100
...
    UsersList:
      title: Users List
      description: |-
        A list of users for many responses.
        It's used to get a list of users for the search, follower and followings lists, banned users list and likes list.
      type: array
      items:
        type: object
        description: User
        properties:
          user: { $ref: "#/components/schemas/User" }
          proPic: { $ref: "#/components/schemas/ImgURL" }
      minItems: 0
      maxItems: 100
...
    OK:
      title: OK
      description: The succesfull response from server.
      type: string
      example: "OK, action successfully completed"
      minLength: 2
      maxLength: 50
    ImgURL:
      title: ImgURL
      description: Image URL specification.
      type: string
      format: url
      #vorrei usare l'estensione webp
      pattern: '^https:\/\/[a-zA-Z.-_\/]*\/users\/[0-9]{5}\/(propic|posts)\/[a-zA-Z]{3}\.jpg$'
      minLength: 34
      maxLength: 128
      example: "https://www.example.com/users/00001/posts/aAk.jpg"
    Timestamp:
      title: Timestamp
      description: The timestamp of the post.
      type: string
      format: date-time
      pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{2}:[0-9]{2}:[0-9]{2}$"
      example: "2020-12-31T23:59:59"
      minLength: 19
      maxLength: 19
  responses:
...
    Unauthorized:
      description: The access token is missing or it's expired.
      content:
        text/plain:
          schema:
            description: Unauthorized
            type: string
            example: "Unauthorized"
            minLength: 12
            maxLength: 30
    #400
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc).
      content:
        text/plain:
          schema:
            description: BadRequest
            type: string
            example: "Bad Request"
            minLength: 11
            maxLength: 30
    #500
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
      content:
        text/plain:
          schema:
            description: InternalServerError
            type: string
            example: "Internal Server Error"
            minLength: 21
            maxLength: 30
    #404
    NotFound:
      description: The resource was not found.
      content:
        text/plain:
          schema:
            description: NotFound
            type: string
            example: "Not Found"
            minLength: 9
            maxLength: 30
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Offset:
      name: offset
      in: query
      description: |-
        The start index of the elements to get. <br>
        <details> 
        <summary> Other details </summary>
        If the value is **not specified**, it will be set to 0. <br>
        If the value is **greater than** the number of elements, an empty array will be returned. <br>
        If the value is **negative**, it will be set to 0. <br>
        If the value is **not an integer**, it will be set to 0.
        </details>
      schema: { type: integer, default: 0, minimum: 0 }
      required: false
    Limit:
      name: limit
      in: query
      description: |-
        The maximum number of elements to get. <br>
        <details> 
        <summary> Other details </summary>
        If the value is **not specified**, it will be set to 10. <br>          
        If the value is **greater than 100**, it will be set to 100. <br>
        If the value is **less than 1**, it will be set to 10. <br>
        If the value is **not an integer**, it will be set to 10. <br>
        If the value is **greater than** the number of elements, it will be set to the total number of elements. 
        </details>
      schema: { type: integer, default: 10, minimum: 1, maximum: 100 }
      required: false